Added SignigAuthorityAssertion Models. Refactored Serializers
