+++ backend/api/fixtures/signing_authority_assertions.json
@@ -1,0 +1,30 @@
+[
+  {
+      "fields": {
+          "description": "I confirm that records evidencing each matter reported under section 11.11 (2) of the Regulation are available on request.",
+          "display_order": "1",
+          "effective_date": "2018-01-01",
+          "expiration_date": null
+      },
+      "model": "api.signingauthorityassertion",
+      "pk": "1"
+  }, {
+    "fields": {
+        "description": "I confirm that I am an officer or employee of the fuel supplier, and that records evidencing my authority to submit this proposal are available on request.",
+        "display_order": "2",
+        "effective_date": "2018-01-01",
+        "expiration_date": null
+    },
+    "model": "api.signingauthorityassertion",
+    "pk": "2"
+  }, {
+    "fields": {
+        "description": "I certify that the information in this report is true and complete to the best of my knowledge and I understand that the Director may require records evidencing the truth of that information.",
+        "display_order": "3",
+        "effective_date": "2018-01-01",
+        "expiration_date": null
+    },
+    "model": "api.signingauthorityassertion",
+    "pk": "3"
+  }
+]
